미션 수정 관련 API 최적화, 상태에 따라 기준시간 변경

+++ src/main/java/com/cofixer/mf/mfcontentapi/dto/req/CreateMissionReq.java
@@ -21,6 +21,6 @@ public class CreateMissionReq {
     /* 미션 시작일(utc) */
     Long startDate;
 
-    /* 기한(일 단위) */
-    Integer deadline;
+    /* 기한(초) */
+    Long deadline;
 }
